requirements.txt
fastapi~=0.115.11
pydantic~=2.10.6
langgraph~=0.3.30
python-dotenv~=1.0.1
pymongo~=3.12.0
langchain-google-genai~=2.1.3
langgraph-checkpoint-sqlite
